% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_export.R
\name{BizDayConfig}
\alias{BizDayConfig}
\title{Create BizDayConfig Object from the zerg::BizDayConfig C++ Class}
\value{
A `BizDayConfig` object from the _C++_ zerg::BizDayConfig Class.
}
\description{
Allows for the creation of a BizDayConfig Object in _C++_ from _R_
using the _C++_ zerg::BizDayConfig class.
}
\examples{
##################
## Constructor

# Construct new FZ::BizDayConfig object called "m"
m = new(FZ::BizDayConfig)


}
