% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expr_split}
\alias{expr_split}
\title{expr_split}
\usage{
expr_split(exprs, N, config_path)
}
\arguments{
\item{exprs}{expr to split}

\item{N}{N to split}

\item{config_path}{meta config file}
}
\description{
expr_split
}
