% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expr_split_to_dt}
\alias{expr_split_to_dt}
\title{expr_split_to_dt}
\usage{
expr_split_to_dt(exprs, delimiter)
}
\arguments{
\item{exprs}{expr to split}

\item{delimiter}{delimiter to split}
}
\description{
expr_split_to_dt
}
