% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_cn_fut_code}
\alias{get_cn_fut_code}
\alias{all_cn_fut_code}
\alias{GetUkey}
\alias{GetUkey2}
\alias{fz_split}
\alias{fz_split_instrument_id}
\alias{fz_now_string}
\alias{fz_now_cob}
\alias{fz_HumanReadableMillisecond}
\alias{fz_replace_time_placeholder}
\alias{fz_mkdirp}
\alias{fz_is_file_existed}
\alias{fz_get_absolute_path}
\alias{fz_dirname}
\alias{fz_basename}
\alias{fz_is_dir}
\alias{fz_read_trading_days}
\alias{fz_path_wildcard}
\alias{str_expand2}
\alias{expr_split}
\alias{expr_split_to_dt}
\title{get_cn_fut_code}
\usage{
get_cn_fut_code(ukeys, file)

all_cn_fut_code(file)

GetUkey(str)

GetUkey2(str, exch)

fz_split(str, separator)

fz_split_instrument_id(str)

fz_now_string()

fz_now_cob()

fz_HumanReadableMillisecond(interval_string)

fz_replace_time_placeholder(str, date)

fz_mkdirp(path)

fz_is_file_existed(path)

fz_get_absolute_path(path)

fz_dirname(path)

fz_basename(path)

fz_is_dir(path)

fz_read_trading_days(path, start = -1L, end = -1L)

fz_path_wildcard(path)

str_expand2(expr, names, values)

expr_split(exprs, N, config_path)

expr_split_to_dt(exprs, delimiter, is_first = TRUE)
}
\arguments{
\item{ukeys}{ukey to look up}

\item{str}{what to work}

\item{exch}{string}

\item{separator}{string used to separate}

\item{interval_string}{what to work}

\item{date}{which date to use}

\item{path}{where to search}

\item{start}{start date}

\item{end}{end date}

\item{expr}{expr to expand}

\item{names}{names to expand}

\item{values}{values to expand}

\item{exprs}{expr to split}

\item{N}{N to split}

\item{config_path}{meta config file}

\item{delimiter}{delimiter to split}

\item{is_first}{true first, false last}
}
\value{
ukey int

ukey int

vector of string

std::vector of string

string

int64_t

uint64_t

string

bool success

bool success

string

string

string

bool success

vector of dates

map <matched, path>

vector of strs
}
\description{
get_cn_fut_code

all_cn_fut_code

GetUkey

GetUkey

split string

SplitInstrumentID

now_string

now_cob

HumanReadableMillisecond like 20us

replace_time_placeholder

mkdirp

IsFileExisted

GetAbsolutePath

Dirname

Basename

IsDir

read_trading_days

path wildcard like /tmp/*.csv

str_expand2

expr_split

expr_split_to_dt
}
