% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_get_absolute_path}
\alias{fz_get_absolute_path}
\title{GetAbsolutePath}
\usage{
fz_get_absolute_path(path)
}
\arguments{
\item{path}{where to work}
}
\value{
string
}
\description{
GetAbsolutePath
}
