% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_HumanReadableMillisecond}
\alias{fz_HumanReadableMillisecond}
\title{HumanReadableMillisecond like 20us}
\usage{
fz_HumanReadableMillisecond(interval_string)
}
\arguments{
\item{interval_string}{what to work}
}
\value{
uint64_t
}
\description{
HumanReadableMillisecond like 20us
}
