% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_path_wildcard}
\alias{fz_path_wildcard}
\title{path wildcard like /tmp/*.csv}
\usage{
fz_path_wildcard(path)
}
\arguments{
\item{path}{where to search}
}
\value{
map <matched, path>
}
\description{
path wildcard like /tmp/*.csv
}
