% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_read_trading_days}
\alias{fz_read_trading_days}
\title{read_trading_days}
\usage{
fz_read_trading_days(path)
}
\arguments{
\item{path}{where to read}
}
\value{
vector of dates
}
\description{
read_trading_days
}
