% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_replace_time_placeholder}
\alias{fz_replace_time_placeholder}
\title{replace_time_placeholder}
\usage{
fz_replace_time_placeholder(str, date)
}
\arguments{
\item{str}{what to work}

\item{date}{which date to use}
}
\value{
string
}
\description{
replace_time_placeholder
}
