% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_split}
\alias{fz_split}
\title{split string}
\usage{
fz_split(str, separator)
}
\arguments{
\item{str}{string to split}

\item{separator}{string used to separate}
}
\value{
vector of string
}
\description{
split string
}
