% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fz_split_instrument_id}
\alias{fz_split_instrument_id}
\title{SplitInstrumentID}
\usage{
fz_split_instrument_id(str)
}
\arguments{
\item{str}{string to split}
}
\value{
std::vector of string
}
\description{
SplitInstrumentID
}
