% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAUtils.R
\name{str_expand}
\alias{str_expand}
\title{expand string}
\usage{
str_expand(str, ..., expand = FALSE)
}
\arguments{
\item{str}{str to be expanded}

\item{...}{list used to expand}

\item{expand}{bool, whether to expand ... list}
}
\description{
expand string
}
