% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_expand2}
\alias{str_expand2}
\title{str_expand2}
\usage{
str_expand2(expr, names, values)
}
\arguments{
\item{expr}{expr to expand}

\item{names}{names to expand}

\item{values}{values to expand}
}
\value{
vector of strs
}
\description{
str_expand2
}
